WIP: Refactor user API - xtor, PlacePiece, ExecuteTurn
I've started looking at it from the user point of view, which is what I
should have done from the start. The user is not going to manipulate
stacks of pieces; they just want to set the inital conditions and then
receive the simulation results.

Refactored the constructor to take board size and number of pieces.

Added PlacePiece() method, taking the place of the AddInFront methods,
which will likely be deleted or at least made private.

ExecuteTurn() is just a stub.

Unit tests are in place for the new methods.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>