Refactor piece manipulation; add ExecuteTurn()
The refactoring of the piece manipulation implementation continues. At this
point it should be almost done. I am still thinking about having the game
board be a list of Piece objects, not just the int IDs, since I do a lot of
results array lookup. Why not just have it all in one place?

Added ExecuteTurn() method. It handles a turn that does not result in the
end of the game; the end-game code is not there yet.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>