Add SimulateOneTurn() and track first/second place
SimulateOneTurn() loops on ExecuteTurn() to determine the probability of
each possible outcome. However, it is not complete, and only works when
simulating one turn. Subsequent turns do not reset the board state, thus
they are invalid.

ExecuteTurn() now tracks the first/second place results in the count
variables in each result record.

I think a redesign is needed to implement the simulation cleanly.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>