Refactor to use Piece object
There is now a Piece class. The arrays of integers (representing the ID
of each piece) have been replaced with using Piece objects. This will
hopefully make the code easier to understand since it is consistent.

Also changed the way the board space list works. Before I had it where
the first piece in the list was on the top. This doesn't seem logical
to me; I don't remember why I did it this way. Now, the first piece in
the list (at index 0) is on the bottom. This way adding a piece to the
space via the Add() method works as expected.

Piece placement is now done in the context of simulating a turn. It is
not done within the PlacePiece method itself. There is also a new array
of Pieces, where the order of the array is the order in which the pieces
are placed. This makes resetting the board for every simulated turn much
easier than it would have been.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>